Rename files and directories for clarity.
